requirements/development.txt
# Include the bare minimum.
-r base.txt

# Debug toolbar
django_debug_toolbar
